requirements.txt
langchain
langchain_community
streamlit
python-dotenv
langchain_google_genai